6-size.c

+++ 0x00-hello_world/6-size.c
@@ -5,10 +5,10 @@
  */
 int main(void)
 {
-	printf("Size of a char: %lu byte(s)\n", sizeof(char));
-	printf("Sizeof of an int: %lu byte(s)", sizeof(char));
-	printf("Sizeof of a long int: %lu byte(s)\n", sizeof(long int));
-	printf("Size of a long long int: %lu byte(s)\n", sizeof(long long int));
-	printf("Size of a float: %lu byte(s)\n", sizeof(float));
-	return (0);
+printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(char));
+printf("Sizeof of an int: %lu byte(s)", (unsigned long)sizeof(char));
+printf("Sizeof of a long int: %lu byte(s)\n", (unsigned long)sizeof(long int));
+printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(long long int));
+printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(float));
+return (0);
 }
